Add Yeoman generated code
? What type of extension do you want to create? New Language Support
Enter the URL (http, https) or the file path of the tmLanguage grammar or press ENTER to s
tart with a new grammar.
? URL or file to import, or none for new: https://raw.githubusercontent.com/textmate/txt2t
ags.tmbundle/master/Syntaxes/Txt2tags.tmLanguage
? What's the name of your extension? txt2tags
? What's the identifier of your extension? txt2tags
? What's the description of your extension? txt2tags language support (t2t)
Enter the id of the language. The id is an identifier and is single, lower-case name such as 'php', 'javascript'
? Language id: txt2tags
Enter the name of the language. The name will be shown in the VS Code editor mode selector.
? Language name: Txt2tags
Enter the file extensions of the language. Use commas to separate multiple entries (e.g. .ruby, .rb)
? File extensions: .t2t
Enter the root scope name of the grammar (e.g. source.ruby)
? Scope names: text.txt2tags